[sksl] Enforce f16 value limits for literals
Besides updating the minimum/maximumValues reported for low precision
floating point types, this uncovered an issue with coercion cost for
float literals. Previously, we were silently preferring half precision
and then would cast back to floats when other parameters selected a
float-typed function instead of a half-typed function.

In order to comply with the new limits and new coercion handling,
some of the SkSL tests have been updated, primarily to use constants
that are within the f16 limit, or to disambiguate generic function
calls.

Bug: b/465408252
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1116397
Auto-Submit: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Thomas Smith <thomsmit@google.com>

+++ tests/sksl/intrinsics/FaceForward.wgsl
@@ -15,7 +15,7 @@ fn _skslMain(xy: vec2<f32>) -> vec4<f32> {
   {
     let _skTemp0 = 1e+30;
     let _skTemp1 = select(-1.0, 1.0, _skTemp0 * 1e+30 < 0);
-    let huge: f32 = f32(_skTemp1);
+    let huge: f32 = _skTemp1;
     let _skTemp2 = vec2<f32>(1.0);
     let _skTemp3 = faceForward(_skTemp2, vec2<f32>(1e+30), vec2<f32>(1e+30));
     let huge2: vec2<f32> = _skTemp3;
